REFACTORING
changed the name of Vector file to Math and added operator functions instead of functions like multiply(). Seperated the code in Shader.h into 2 well organized files; Shader and Mesh; both in .cpp and have a header file. Now added the ability to automatically get the indices from a texture inorder to save up computing power. Inaddition, added the functionality that  the mesh we have has the same number of vertices as that of the texture size(in pixels). All around cut the bloat off and made quality of life improvements.