Object buffers and Move semantics
Improved the *Texture* class by adding copy and move constructors so when a *Texture* is used to create a *Mesh*, no 2 instances share the same pointer data, this improves stability and reliability. Moreover, improved the *Mesh* class so that includes all needed *Textures* of an object and the constructor now uses move semantics to take advantage of the new improved *Texture*. Moreover, the *Shader* class now can bind to all buffers and texture data that a *Mesh* class has and directly draws it, hence creating an VAO behaviour without having the need to have multiple vertex arrays binded and unbined to GL_BUFFER; instead, the *Shader* class binds the buffers and textures draws it and moves on to the next object and does the same. In addition, it can check if said buffers and textures have already been generated or not, making things more automated.