More GPU usage
Improved the *Point_Cloud* class such that it can now extract data faster. Added *mat3* and *mat4* type hashing functions to *Math.h* file and added a function to turn *mat3* and *mat4* types to std::vector. Moreover, moved the calculation of transformation matrices from CPU to GPU. Instead of using the various *vec3* and *float* types to calculate the matrices inside of *init_matrices* functions, all of the required members are now uniforms and are used in the shader to calculate the matrices. This should make the GPU more effective and less stressful for the CPU. Moreover, removed the various *graphics_XX_container* structs and their functions for a more open aproach when it comes to adding uniforms and updating them. Now am uisng various std::unordered_map members inside the shader which can be easily used with their new functions to retrieve the uniforms we are using and update them with alot more automatic approach than that of the *graphics_XX_container* types. Hence, now adding a uniform to any of the maps is suffice for it to be correctly updated without changing any old parameters or refactoring.