LAS optimization
optimized the *Point_Cloud* class that now the data extraction is faster, also added the capability to extract colors when we extract the point data. Optimised some constructors and the *extract_from_LAS_file* function inside the *Mesh* class so everything is now a bit faster and more readable. Moreover, added the capability that the an instance of *Texture* class can be empty(as in no actual byte size of anything it just points to NULL). In addition added checks inside the *bind_mesh_buffer_and_textures* function in the *Shader* class that checks if a texture is NULL or if a buffer is empty before it attempts to bind, which makes the binding process faster. Optimized LAS_POINTS shader to not use useless features regarding rendering from data extracted from LAS, such features are light, normal mapping... Before said optimizations the time needed for the screen to first display something from a LAS with 80M points without color extraction was 1.28 minutes, now it needs 18 seconds(color extraction included).